New Q design first draft.
A QueueCapability is used to create queues and submit them to an ExecutionEngine (there is only one per resource).

ExecutionEngine executes a transaction (tx).
ExecutionEngine defines a TransactionLogic and runs it against tx.
It basically executes begin method and delegates to a ITransactionCoordinator the decision of commiting or aborting tx).
This execution is done when IExecutionEngine.submit(ITransaction tx) is called.

If ICompositeExecutionEngine.submit(ITransaction tx) is called, instead, execution engine waits for calls to begin, commir or abort,
for applying these action in the tx.
This behaviour allows the construction of two phase commit transactions, where tx is a component of the composite transaction.
This behaviour allows creation of a transaction of transactions (compositeTx), and is it replicable, allowing constructions in the form of a
compositeTx of compositeTxs.

There are describing desired behaviour for a suitable implementation.
However they are not completed.