Add progress bar to delete link in the menu.
When deleting a network confirmation is required.
After confirmation is given, a progress bar is now displayed.

There was a problem with the progress bar being frozen, when calling
window.location.href.
Thus, this call has been replaced by creating a custom form which
loads desired href when it is submitted.

TODO test progress bar is not frozen now.